moving EventHubsUtils to eventhubs.common - going to add constants there and soon we'll support Spark core
